requirements.txt
pony==0.7.16
uvicorn==0.18.3
fastapi==0.85.0
PyJWT==1.7.0
coverage==6.5.0
fastapi-mail==1.2.0